Move Reports folder to Core.
