+++ GitHubBugReport.Core/GitHubBugReport.Core.csproj
@@ -67,6 +67,21 @@
     <Compile Include="Query\SyntaxParser.cs" />
     <Compile Include="Query\Token.cs" />
     <Compile Include="Query\Tokenizer.cs" />
+    <Compile Include="Reports\Alert.cs" />
+    <Compile Include="Reports\Config.cs" />
+    <Compile Include="Reports\CsvReport\CsvTableReport.cs" />
+    <Compile Include="Reports\CsvReport\CsvWriter.cs" />
+    <Compile Include="Reports\EmailReport\AlertReport.cs" />
+    <Compile Include="Reports\EmailReport\AlertReport_Diff.cs" />
+    <Compile Include="Reports\EmailReport\AlertReport_NeedsMSResponse.cs" />
+    <Compile Include="Reports\EmailReport\AlertReport_Untriaged.cs" />
+    <Compile Include="Reports\ExpressionUntriaged.cs" />
+    <Compile Include="Reports\HtmlReport\HtmlQueryCountLink.cs" />
+    <Compile Include="Reports\HtmlReport\HtmlTableReport.cs" />
+    <Compile Include="Reports\HtmlReport\QueryReport.cs" />
+    <Compile Include="Reports\IssueEntry.cs" />
+    <Compile Include="Reports\TableReport.cs" />
+    <Compile Include="Reports\XlsxReport\HistoryReport.cs" />
     <Compile Include="Repositories\Models\Repository.cs" />
     <Compile Include="Repositories\Services\IRepositoryService.cs" />
     <Compile Include="Repositories\Services\OctoKitRepositoryService.cs" />
