+++ GitHubBugReport.Core/Reports/EmailReport/AlertReport_NeedsMSResponse.cs
@@ -2,11 +2,10 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Web;
 using GitHubBugReport.Core.Issues.Models;
 using GitHubBugReport.Core.Util;
 
-namespace BugReport.Reports.EmailReports
+namespace GitHubBugReport.Core.Reports.EmailReport
 {
     public class AlertReport_NeedsResponse
     {
