+++ GitHubBugReport.Core/Reports/HtmlReport/HtmlQueryCountLink.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
+﻿using System.Collections.Generic;
 using System.Linq;
 using GitHubBugReport.Core.Query;
 
-namespace BugReport.Reports
+namespace GitHubBugReport.Core.Reports.HtmlReport
 {
     public class HtmlQueryCountLink
     {
