+++ GitHubBugReport.Core/Reports/HtmlReport/QueryReport.cs
@@ -3,10 +3,9 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
-using System.Web;
 using GitHubBugReport.Core.DataModel;
 
-namespace BugReport.Reports
+namespace GitHubBugReport.Core.Reports.HtmlReport
 {
     public class QueryReport
     {
